[BUGFIX] Date fields with hour "00" get deleted on saving
When entering a date/time like "00:00 02-04-1975" in a date widget
(or choosing this date/time using the datepicker), the date would get
discarded on saving.

With this commit, the date now is only discarded if it is an empty string.

This commit also cleans up the corresponding methods a bit.