[BUGFIX] Generate JavaScript in `typo3temp/assets/mkforms/`
Generating JavaScript files in `typo3temp/assets/mkforms/` instead of in
`typo3temp/mkforms/` makes sure that the generated files can be loaded in
installations that make use of the `helhum/typo3-secure-web` package
(that symlinks `typo3temp/assets/` into the document root, but does not
symlink `typo3temp/` itself.

This is the second part of the fix for #106.